+++ plugins/DNNCommunity.DNNDocs.Plugins/DNNCommunity.DNNDocs.Plugins.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DNNCommunity.DNNDocs.Plugins.GitHubStats", "DNNCommunity.DNNDocs.Plugins.GitHubStats.csproj", "{FA0B4F79-66FD-4F55-BBB9-0CAC9CD4F095}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DNNCommunity.DNNDocs.Plugins", "DNNCommunity.DNNDocs.Plugins.csproj", "{FA0B4F79-66FD-4F55-BBB9-0CAC9CD4F095}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
