+++ plugins/DNNCommunity.DNNDocs.Plugins/MoreLinksProcessor.cs
@@ -2,13 +2,11 @@
 using System.Composition;
 using System.Collections.Immutable;
 
-namespace DNNCommunity.DNNDocs.Plugins.GitHubStats
+namespace DNNCommunity.DNNDocs.Plugins
 {
-    [Export(nameof(PageStatsProcessor), typeof(IPostProcessor))]
-    public class PageStatsProcessor : IPostProcessor
+    [Export(nameof(MoreLinksProcessor), typeof(IPostProcessor))]
+    public class MoreLinksProcessor : IPostProcessor
     {
-        // TODO: implements IPostProcessor
-
         public ImmutableDictionary<string, object> PrepareMetadata(ImmutableDictionary<string, object> metadata)
         {
             // TODO: add/remove/update property from global metadata
