+++ plugins/DNNCommunity.DNNDocs.Plugins/PageStatsProcessor.cs
@@ -2,10 +2,10 @@
 using System.Composition;
 using System.Collections.Immutable;
 
-namespace DNNCommunity.DNNDocs.Plugins.GitHubStats
+namespace DNNCommunity.DNNDocs.Plugins
 {
-    [Export(nameof(RepoStatsProcessor), typeof(IPostProcessor))]
-    public class RepoStatsProcessor : IPostProcessor
+    [Export(nameof(PageStatsProcessor), typeof(IPostProcessor))]
+    public class PageStatsProcessor : IPostProcessor
     {
         // TODO: implements IPostProcessor
