+++ plugins/DNNCommunity.DNNDocs.Plugins/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("DNNCommunity.DNNDocs.Plugins.GitHubStats")]
+[assembly: AssemblyTitle("DNNCommunity.DNNDocs.Plugins")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("DNNCommunity.DNNDocs.Plugins.GitHubStats")]
+[assembly: AssemblyProduct("DNNCommunity.DNNDocs.Plugins")]
 [assembly: AssemblyCopyright("Copyright ©  2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
