+++ plugins/DNNCommunity.DNNDocs.Plugins/app.config
