+++ plugins/DNNCommunity.DNNDocs.Plugins/packages.config
